requirements.txt
fastapi
uvicorn[standard]
celery
redis
playwright
beautifulsoup4
jinja2
python-multipart
requests
aiofiles
Pillow
csscompressor
